+++ tests/testthat/test-lav_betaselect_mod_mg.R
@@ -12,24 +12,40 @@ library(lavaan)
 dat <- data_test_medmod
 dat$iv_mod <- dat$iv * dat$mod
 head(dat)
+dat$gp <- rep(c("gp2", "gp1"), size = nrow(dat) / 2)
 
-dat_c <- scale(dat,
+dat1_c <- scale(dat[dat$gp == "gp2", -8],
                center = TRUE,
                scale = FALSE)
-dat_c <- as.data.frame(dat_c)
+dat2_c <- scale(dat[dat$gp == "gp1", -8],
+               center = TRUE,
+               scale = FALSE)
+dat1_c <- as.data.frame(dat1_c)
+dat2_c <- as.data.frame(dat2_c)
+dat1_c$gp <- "gp2"
+dat2_c$gp <- "gp1"
+dat_c <- rbind(dat1_c, dat2_c)
 dat_c$iv_mod <- dat_c$iv * dat_c$mod
-colMeans(dat_c)
-apply(dat_c,
+colMeans(dat_c[, -8])
+
+apply(dat_c[, -8],
       MARGIN = 2,
       sd)
 
-dat_z <- scale(dat,
+dat1_z <- scale(dat[dat$gp == "gp2", -8],
                center = TRUE,
                scale = TRUE)
-dat_z <- as.data.frame(dat_z)
+dat2_z <- scale(dat[dat$gp == "gp1", -8],
+               center = TRUE,
+               scale = TRUE)
+dat1_z <- as.data.frame(dat1_z)
+dat2_z <- as.data.frame(dat2_z)
+dat1_z$gp <- "gp2"
+dat2_z$gp <- "gp1"
+dat_z <- rbind(dat1_z, dat2_z)
 dat_z$iv_mod <- dat_z$iv * dat_z$mod
-colMeans(dat_z)
-apply(dat_z,
+colMeans(dat_z[, -8])
+apply(dat_z[, -8],
       MARGIN = 2,
       sd)
 
@@ -40,13 +56,16 @@ dv ~ med + cov2
 "
 fit <- sem(mod,
            data = dat,
-           likelihood = "wishart")
+           likelihood = "wishart",
+           group = "gp")
 fit_c <- sem(mod,
              data = dat_c,
-             likelihood = "wishart")
+             likelihood = "wishart",
+             group = "gp")
 fit_z <- sem(mod,
              data = dat_z,
-             likelihood = "wishart")
+             likelihood = "wishart",
+             group = "gp")
 
 std <- standardizedSolution(fit)
 std <- setNames(std$est.std, lav_partable_labels(std))
@@ -59,17 +78,18 @@ test_that("Check variables are centered", {
   prods <- find_all_products(fit_c,
                              parallel = FALSE,
                              progress = FALSE)
-  expect_false(check_centered(fit,
-                              prods = prods))
-  expect_true(check_centered(fit_c,
-                             prods = prods))
-  expect_true(check_centered(fit_z,
-                             prods = prods))
+  expect_true(all(!check_centered(fit,
+                              prods = prods)))
+  expect_true(all(check_centered(fit_c,
+                             prods = prods)))
+  expect_true(all(check_centered(fit_z,
+                             prods = prods)))
 })
 
 test_that("Coefficient of the component terms", {
 
-  chk_names <- c("med~mod", "med~iv", "med~iv_mod")
+  chk_names <- c("med~mod", "med~iv", "med~iv_mod",
+                 "med~mod.g2", "med~iv.g2", "med~iv_mod.g2")
 
   coef(fit)[chk_names]
   coef(fit_c)[chk_names]
@@ -94,7 +114,7 @@ test_that("Coefficient of the component terms", {
 
 
   # Should be equal
-  expect_equal(coef(out_c)[chk_names],
-               coef(fit_z)[chk_names])
+  expect_equal(coef(out)[chk_names],
+               coef(out_c)[chk_names])
 
 })